candib0t-v1/pyproject.toml
[tool.poetry]
name = "disc0rd"
version = "0.1.0"
description = ""
authors = ["b <b@elastic.ventures>"]

[tool.poetry.dependencies]
python = "^3.10"
discord = "^1.7.3"
python-dotenv = "^0.20.0"
"discord.py" = "^1.7.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
disc0rd = "disc0rd.bot:cli"
hn = "hackern3ws.scraper:cli"
